feat: 한 종류의 옵션에 여러가지 제품을 넣을 수 있도록 수정

+++ services/items.services.js
@@ -1,6 +1,5 @@
 const ItemRepository = require('../repositories/items.repositories');
 const { CustomError, ServiceReturn } = require('../customClass');
-const { checkPrimeSync } = require('crypto');
 
 class ItemService {
     itemRepository = new ItemRepository();
@@ -8,13 +7,14 @@ class ItemService {
     getItem = async (item_id) => {
         const findItemId = await this.itemRepository.itemFindOne(item_id);
         if (!findItemId) throw new CustomError('해당하는 아이템은 존재하지 않습니다.', 403);
+        console.log(findItemId);
         const findOption = await this.itemRepository.optionFindOne({ option_id: findItemId.option_id });
         const Item = {
-            item_id: findOption.item_id,
-            name: findOption.Items[0].name,
-            price: findOption.Items[0].price,
-            type: findOption.Items[0].type,
-            amount: findOption.Items[0].amount,
+            item_id: findItemId.item_id,
+            name: findItemId.name,
+            price: findItemId.price,
+            type: findItemId.type,
+            amount: findItemId.amount,
             extra_price: findOption.extra_price,
             shot_price: findOption.shot_price,
             ice: findOption.ice,
@@ -36,12 +36,15 @@ class ItemService {
         return new ServiceReturn('상품 리스트를 성공적으로 불러왔습니다.', 200, itemList);
     };
 
-    createItem = async ({ name, price, type, extra_price, shot_price, ice }) => {
+    createItem = async ({ name, price, type }) => {
         const overlapName = await this.itemRepository.itemFindOne({ name: name });
         if (overlapName) throw new CustomError('같은 이름의 상품이 존재합니다.', 403);
         if ((!name, !price, !type)) throw new CustomError('필수 입력 항목이 충족되지 않았습니다.', 403);
+        let option_id = 0;
+        if (type === 'coffee') option_id = 1;
+        if (type === 'non-coffee') option_id = 2;
 
-        await this.itemRepository.createItem({ name, price, type, extra_price, shot_price, ice });
+        await this.itemRepository.createItem({ name, price, type, option_id });
 
         return new ServiceReturn('상품 등록이 정상적으로 완료되었습니다.', 200);
     };
@@ -58,14 +61,16 @@ class ItemService {
         return new ServiceReturn('상품 수정이 정상적으로 완료되었습니다.', 200);
     };
 
-    deleteItem = async ({ item_id }) => {
+    deleteItemcheck = async ({ item_id, answer }) => {
         const findItemId = await this.itemRepository.itemFindOne({
             item_id: item_id,
         });
         if (!findItemId) throw new CustomError('해당하는 아이템은 존재하지 않습니다.', 403);
 
-        await this.itemRepository.deleteItem({ item_id, option_id: findItemId.option_id });
+        if (findItemId.amount !== 0 && answer === '예') return await this.itemRepository.deleteItem({ item_id, option_id: findItemId.option_id });
+        if (findItemId.amount !== 0) return new ServiceReturn('삭제하려는 상품의 재고가 남아있습니다. 정말로 삭제하시겠습니까?', 200, findItemId);
 
+        await this.itemRepository.deleteItem({ item_id, option_id: findItemId.option_id });
         return new ServiceReturn('상품 삭제가 정상적으로 완료되었습니다.', 200);
     };
 }
